# api/static/swagger.yaml
swagger: "2.0"
info:
  title: "MLP API"
  description: "API Guide for accessing MLP API"
  version: "0.4.0"
host: "localhost:8080"
basePath: "/v1"
...
schemes:
  - "http"
paths:
  "/projects":
    get:
      tags: ["project"]
      summary: "List existing projects"
      description: "Projects can be filtered by optional `name` parameter"
      parameters:
        - in: "query"
          name: "name"
          required: false
          type: "string"
      responses:
        200:
          description: "OK"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Project"
    post:
      tags: ["project"]
      summary: "Create new project"
      parameters:
        - in: "body"
          name: "body"
          description: "Project object that has to be added"
          required: true
          schema:
            $ref: "#/definitions/Project"
      responses:
        201:
          description: "Created"
          schema:
            $ref: "#/definitions/Project"
        400:
          description: "Invalid request format"
        409:
          description: "Project with the same name already exists"
...
definitions:
  Project:
    type: "object"
    required:
      - name
    properties:
      id:
        type: "integer"
        format: "int32"
      name:
        type: "string"
      mlflow_tracking_url:
        type: "string"
      administrators:
        type: "array"
        items:
          type: "string"
      readers:
        type: "array"
        items:
          type: "string"
      team:
        type: "string"
      stream:
        type: "string"
      labels:
        type: "array"
        items:
          $ref: "#/definitions/Label"
      created_at:
        type: "string"
        format: "date-time"
      updated_at:
        type: "string"
        format: "date-time"
  Label:
    type: "object"
    properties:
      key:
        type: "string"
      value:
        type: "string"
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
security:
  - Bearer: []